requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.1
pydantic-settings>=2.1.0
langgraph>=0.0.26
langchain-core>=0.1.27
langchain-groq>=0.0.1
chromadb>=0.4.18
pypdf>=3.17.1
pandas>=2.1.3
numpy>=1.26.2
pytest>=7.4.3
httpx>=0.25.2
python-multipart>=0.0.6
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0
groq>=0.4.0
tiktoken>=0.5.1
langchain-huggingface>=0.0.8
transformers>=4.37.2
torch>=2.2.0
huggingface-hub>=0.21.3
